removed legacy UI and made changes to support the project scoped integrations and app config
