+++ apps/web/src/app/[projectId]/ai/page.tsx
