+++ apps/web/src/app/[projectId]/routes/page.tsx
@@ -1,0 +1,28 @@
+"use client";
+import React from "react";
+import { Stack, Text, Group } from "@mantine/core";
+import RoutesPanel from "@/components/panels/routesPanel";
+import RouterFilter from "@/components/filters/routerFilter";
+import RouterPagination from "@/components/filters/routerPagination";
+import { useParams } from "next/navigation";
+
+const ProjectRoutesPage = () => {
+	const { projectId } = useParams();
+
+	return (
+		<Stack style={{ height: "100%" }} p="md" gap="md">
+			<Group justify="end" align="center">
+				<Group>
+					<RouterPagination />
+					<RouterFilter />
+				</Group>
+			</Group>
+
+			<div style={{ flex: 1, overflowY: "auto" }}>
+				<RoutesPanel projectId={projectId as string} />
+			</div>
+		</Stack>
+	);
+};
+
+export default ProjectRoutesPage;
