+++ apps/web/src/app/editor/page.tsx
@@ -2,7 +2,7 @@
 import { redirect } from "next/navigation";
 
 const page = async () => {
-  return redirect("/");
+	return redirect("/routes");
 };
 
 export default page;
